feat: Simple splitting of modules,add debug/go image build
