+++ lua/docker/debugGoDockerfile
@@ -1,0 +1,16 @@
+FROM golang:1.19.5 as golang
+
+FROM base/image:latest
+
+COPY --from=golang /usr/local/go /root/go
+
+ENV PATH="${PATH}:/root/go/bin"
+
+RUN export PATH
+
+RUN go env -w GOPROXY="https://goproxy.cn,direct"
+
+RUN go install github.com/go-delve/delve/cmd/dlv@latest
+
+CMD ["dlv","dap","-l","0.0.0.0:38697","--log","--log-output=dap"]
+
